Atom(input,button)作成
