+++ src/components/atoms/button/PrimaryButton.tsx
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+export const PrimaryButton = (props) => {
+  const { children, onClick } = props;
+  return <SButton onClick={onClick}>{children}</SButton>;
+};
+
+const SButton = styled.button`
+  background-color: #40514e;
+`;
