+++ src/components/atoms/input/Check.tsx
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+export const Check = (props) => {
+  const { type,onClick,placeholder } = props;
+  return <SInput type={type} onClick={onClick} placeholder={placeholder} />;
+};
+
+const SInput = styled.input`
+
+`;
